requirements.txt
langchain-groq==1.1.2
Flask==3.1.3
groq==0.37.1
langchain==1.2.10
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.14
langchain-text-splitters==1.1.1
langgraph==1.0.9
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
python-dotenv==1.2.1
Werkzeug==3.1.6